Add integer value support to shader data types. Convert Renderer2Ds texture index from float to unsigned integer.

+++ pine/include/Pine/Renderer/Renderer2D.hpp
@@ -36,7 +36,7 @@ struct QuadVertex
     Vec3 Position = {};
     Vec4 Color = {};
     Vec2 TextureCoord = {};
-    float TextureIndex = {};
+    uint32_t TextureIndex = {};
     float TilingFactor = {};
 };
